refactor(payments): change amount to decimal in Payment entity
- Updated `amount` column type from `int` to `decimal` in `Payment` entity.
- Ensures higher precision for financial transactions.

+++ src/infra/entities/payment.ts
@@ -34,7 +34,7 @@ export class Payment {
   @ForeignKey<Order>('Order', { onDelete: 'CASCADE', onUpdate: 'CASCADE' })
   orderId: string
 
-  @Column('int')
+  @Column('decimal')
   amount: number
 
   @Column({
